Move errors to pkg/utils and move cli to cmd/cli
